+++ CHANGELOG.md
@@ -5,8 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-### Unreleased
+## Unreleased
 
+### Added
+
+- Added ton lite client network provider
+- Added tolk verifier
+- Wallet v4 extended support: added v4r1, v4 is treated as v4r2
 - Added possibility to specify custom `manifestUrl` in blueprint configuration
 
 ## [0.34.0] - 2025-05-20