+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added possibility to specify custom `manifestUrl` in blueprint configuration
 - Added documentation about the [Tact plugin by TON Studio](https://plugins.jetbrains.com/plugin/27290-tact)
 
+### Changed
+
+- Updated documentation about Tact wrappers
+
 ## [0.34.0] - 2025-05-20
 
 ### Added